feat(ux): Move check for updates into settings screen
