+++ src/auth/screens/user-options.screen.js
@@ -1,11 +1,19 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { ScrollView, StyleSheet, FlatList } from 'react-native';
+import {
+  ScrollView,
+  StyleSheet,
+  FlatList,
+  Text,
+  TouchableOpacity,
+} from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { NavigationActions } from 'react-navigation';
+import { version } from 'package.json';
+import codePush from 'react-native-code-push';
 
 import { ViewContainer, SectionList } from 'components';
-import { colors, fonts } from 'config';
+import { colors, fonts, normalize } from 'config';
 import { openURLInView, resetNavigationTo, translate } from 'utils';
 import { signOut, changeLanguage } from 'auth';
 import languages from './language-settings';
@@ -32,6 +40,26 @@ const styles = StyleSheet.create({
     color: colors.red,
     ...fonts.fontPrimary,
   },
+  update: {
+    flex: 1,
+    alignItems: 'center',
+    marginVertical: 40,
+  },
+  updateText: {
+    color: colors.greyDark,
+    ...fonts.fontPrimary,
+  },
+  updateTextSub: {
+    fontSize: normalize(11),
+  },
+});
+
+const updateText = lang => ({
+  check: translate('auth.profile.codePushCheck', lang),
+  checking: translate('auth.profile.codePushChecking', lang),
+  updated: translate('auth.profile.codePushUpdated', lang),
+  available: translate('auth.profile.codePushAvailable', lang),
+  notApplicable: translate('auth.profile.codePushNotApplicable', lang),
 });
 
 class UserOptions extends Component {
@@ -42,8 +70,20 @@ class UserOptions extends Component {
     navigation: Object,
   };
 
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      updateText: updateText(props.language).check,
+    };
+  }
+
   componentWillReceiveProps(nextState) {
     if (nextState.language !== this.props.language) {
+      this.setState({
+        updateText: updateText(nextState.language).check,
+      });
+
       const navigationParams = NavigationActions.setParams({
         params: {
           title: translate('auth.userOptions.title', nextState.language),
@@ -54,6 +94,28 @@ class UserOptions extends Component {
       nextState.navigation.dispatch(navigationParams);
     }
   }
+
+  checkForUpdate = () => {
+    if (__DEV__) {
+      this.setState({
+        updateText: updateText(this.props.language).notApplicable,
+      });
+    } else {
+      this.setState({ updateText: updateText(this.props.language).checking });
+      codePush
+        .sync({
+          updateDialog: true,
+          installMode: codePush.InstallMode.IMMEDIATE,
+        })
+        .then(update => {
+          this.setState({
+            updateText: update
+              ? updateText(this.props.language).available
+              : updateText(this.props.language).updated,
+          });
+        });
+    }
+  };
 
   signOutUser() {
     const { signOutByDispatch, navigation } = this.props;
@@ -109,6 +171,15 @@ class UserOptions extends Component {
               underlayColor={colors.greyLight}
             />
           </SectionList>
+
+          <TouchableOpacity style={styles.update} onPress={this.checkForUpdate}>
+            <Text style={styles.updateText}>
+              GitPoint v{version}
+            </Text>
+            <Text style={[styles.updateText, styles.updateTextSub]}>
+              {this.state.updateText}
+            </Text>
+          </TouchableOpacity>
         </ScrollView>
       </ViewContainer>
     );
